requirements.txt
fastapi
uvicorn
python-fasthtml
google-generativeai
python-dotenv
requests
pydantic
typing-extensions
python-multipart